move controls to new file, shader to holoplay core
